+++ src/ansys/additive/core/porosity.py
@@ -20,6 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 """Provides input and result summary containers for porosity simulations."""
+import math
+
 from ansys.api.additive.v0.additive_domain_pb2 import PorosityInput as PorosityInputMessage
 from ansys.api.additive.v0.additive_domain_pb2 import PorosityResult
 from ansys.api.additive.v0.additive_simulation_pb2 import SimulationRequest
@@ -80,6 +82,8 @@ class PorosityInput:
     def __validate_range(self, value, min, max, name):
         if value < min or value > max:
             raise ValueError("{} must be between {} and {}.".format(name, min, max))
+        elif math.isnan(value):
+            raise ValueError("{} must be a number.".format(name))
 
     @property
     def id(self) -> str: